Use a Python fetcher to get HTML from live forum.
This is a workaround for the difficulty of getting the correct HTML based on the
database contents. When using own bbcode parser, I was getting wrong results.
Most painfully, broken/mangled links.

This change makes the archiver work much much slower, unfortunately. But I've
researched other options and there really isn't a better option. bbcode parsing
is so intricate and so tied to the rest of the forum, that I would have to
reimplement part of phpBB here. Reusing phpBB parsing code is also a pain in the
neck, because convert.php would have to become very tightly coupled with phpBB
code, and I didn't want that either. Fetch forum via HTTP and extracting post
content from HTML sadly is the best option I could find.

Related threads:

-  Somebody asking how to display a post on another page:
   https://area51.phpbb.com/phpBB/viewtopic.php?t=31746
-  phpBB wiki pages
   -  https://wiki.phpbb.com/Function.generate_text_for_display
      Unfortunately this requires pulling in half of phpBB
   -  https://wiki.phpbb.com/Tutorial.Parsing_text
-  A blog entry about displaying posts and topics on external pages
   https://blog.phpbb.com/2009/11/09/how-to-display-posts-and-topics-on-external-pages/

Fixes #2.